Switch answer validator tests to use FakeFS
Rather than stubbing the results of `Data.load`, which makes these very
tied to the implementation and the particular order that different
`Data.load` calls are made.

Also switched use of a single `configure_data` `let` block with keys for
the answers and configure data to multiple `let` blocks for each of
these; I can't see why we would want to use one big hash rather than do
this.