Use lazy loaded methods rather than setters
These methods were slightly confusing previously, as they look like
setters (`something=`) and take a single argument, but then they do not
actually set the value to the argument or a slight permutation of it,
rather they use the passed argument to determine what the value to be
set should be. Using lazy loaded methods instead achieves the same
purpose slightly more obviously.