Use FileSystem rather than FakeFS to run tests
With the previous commit various tests which did not previously use
`FakeFS` were failing intermittently, I think due to certain files they
implicitly depended on not being set up correctly. The failures were
likely intermittent as we used `FakeFS` directly, which does not clear
out the file system before or after it is activated, so depending on
which tests ran previously needed files may or may not have existed in
the fake filesystem.

Switching to use our `FileSystem` wrapper class fixes this as it sets up
the file system afresh for each test as Metalware expects.