Load validation file through dedicated class
Now that their is significant use of validated files, it is easier
if the validation occurs when the file is loaded. To acheive this
a dedicated Validator::Loader class has been created. Currently it
has load methods that load configure.yaml and answer files.

It is possible that in the future it is expanded to include save
file methods and create a data cache.