Determine IP on configured build interface
Previously we determined the IP of the Metalware deployment server on
the build interface by relying on a function from Clusterware and the
server being a Controller appliance; if one of these pre-conditions
failed then we fell back to using `hostname -i` which could sometimes
give the incorrect answer.

This commit changes this so we always use a function from Clusterware,
which has just been copied in to remove the dependency on Clusterware;
we determine the build interface to use based on the newly rendered
`server.yaml` file, which defines properties of the deployment server
itself. This should make this more robust and configurable.