Update answers and cache when removing a group
The `remove group` command has been renamed `remove primary_group`
so that its named closer to its intended purpose.

When removing a primary group, the answer files for that group and
and nodes part of the primary group are deleted. Nodes that have it
as a secondary group are unaffected, so it will still remain as a
valid "group".

Currently when a group is removed, it is deleted from the cache. This
does currently have the affect of altering the group_index of any
subsequent primary groups