Expand the unified Validator::Loader class
As both the load and save methods in the Validator::Loader need to
use the same paths, they have been included in the base class. This
is implemented using a find_path method.

Additionally, loading answer files is now supported. The group_cache
file can now be saved through the Saver. The syntax of the saver has
been adjusted by using a SaveHelper class that takes the data to be
saved and then waits to be told how to save it.

Previously Validator::Answers would take a relative path and determine
the absolute path and section from that. However this adds complexity
and is being deprecated. However it can not be removed until Dependency
is switched over to use the file loader. Until such time, backwards
compatability must be maintained.