Improve testing of remove primary_group
Switched the test over to use `nodeattr` with a mocked genders file
instead of stubbing the entire method.

Added additional assertions to check that other nodes still belong
to the correct groups after a primary_group has been deleted.

Fixed a bug in NodeattrInterface in which a rouge space in the genders
file will causes primary_groups to fail

Also fixed rubocop exceptions. Their are two remaing violations which
will be fixed at a latter date. One is Validator::Answers is two long.
The other is Configurator has to many inputs