Have build templates defaults not set at CLI level
We want to add the ability to specify the PXELINUX and Kickstart
templates used when rendering in `build` within the repo config files
themselves (see https://github.com/alces-software/metalware/issues/143).
This means the new order of precedence for these templates should be:
value specified by option passed > value specified in repo for node >
default value.

For several reasons this means once this has been done we should no
longer be specifying the default template within the CLI config to be
used as the Commander default:

- firstly this would be misleading to users, to have a default specified
in the help which may not be used even if no option is passed (if a repo
value is used instead);

- there is also no way that we could show the actual value which will be
used, taking into account the repo value in the CLI help, as we cannot
know what this will be when forming the help without parsing the whole
config, and even if we could do this the value from the repo could
differ between different nodes in a group;

- finally there is the practical reason that, if we left the current
values of `default` for the template names specified in the CLI config,
by the time we got this value within the `build` command we would have
no way of knowing if we have this value because it was passed to us
explicitly using an option or we were just being given the default;
therefore we would have know way of knowing if the value from the repo
should override this.