Update Validator::Answer to use Loader methods
To make the validators more constistant, Validator::Answer no-longer
excepts a relative path. This removes the need to workout which
answer file it is validating. It also requires the section as an
input which also reduces the code complexity.

As the Validator::Loader is integrated with Config, it has also been
removed as an input. The `configure.yaml` file is always loaded
through the Loader and thus is validated.