Fix Duplicated node bug in hosts file
Nodes can be duplicated in the hosts file if they belong to an
additional gender that is also a primary group. This is because
nodeattr returns all the nodes belong to a gender even if they
do not belong to the primary group.

This commit makes the `nodes` method in the GroupNamespace only
return nodes in the primary group using the new NodeattrInterface
method: `nodes_in_primary_group`