Implement `basic` build method (closes #145)
This provides an alternative to what we did previously, where
`kickstart` and `pxelinux` templates were always rendered when building
a node.

Instead, by specifying `build_method: basic` within the repo config for
a node, only the build `files` and single `basic` template will be
rendered for that node when it is built. It is then up to the user to
ensure this rendered file is run and, as for the Kickstart template, up
to the repo author to ensure this template includes reporting back to
Metalware when the node has built itself.

This build method is intended as a basic way for Metalware to support
setting up the Metalware deployment server itself; it also makes
building non-PXE booting nodes, such as cloud nodes, simpler using
Metalware.