Switch data structure used in GroupCache
Previously the groups where stored as an array where their index is
the position in the array. However as we want static group indices
even when a value is removed from the array, the data structure has
to change.

Now the next index to be used is stored in the yaml file and is
incremented each time a group is added. The groups are now stored in
a hash where the group name is the key and the index stored as the
value.

When a group is removed, it is deleted from the has but the next index
remains untouched