Implement templater method to render managed files
This render templates to a file where only part of the file is managed
by Metalware:

- if the file does not exist yet, it will be created with a new managed
section;

- if it exists without a managed section, the new section will be
appended to the bottom of the current file;

- if it exists with a managed section, this section will be replaced
with the new managed section.

This will soon be used when we render the `genders` and `hosts` files,
to allow admins to make one-off edits to these files without needing to
care about Metalware, while still allowing future Metalware commands to
update these files.