Update group_spec to use the new GroupCache
Previously group_spec would manually create it's own cache files
and then tests the configurator against them. However this is now
been replaced using the new GroupCache.

A primary_group method had to be made public in the GroupCache so
the tests could access a list of nodes as a string. This method
was previously private with a slightly different name.

Also as the Configurator and Rspec both have a copy of the GroupCache,
the two do not play well together as the save method in one Cache
doesn't trigger a reload in the other. This is been fixed by adding
a force_reload flag which can be used by the tester.