Memoize Node#repo_config
Forming the `repo_config` for a node is a slow operation, and it
shouldn't change between the different situations this method will be
used, therefore this should be safe to do and speed things up a bit.

Note: the repo config for a node will sometimes change within a single
run of a Metalware command, such as when we pass a different value for
`firstboot` in when running `Build`, but this change shouldn't effect
that as we go directly through the `Templater` in that case.