Extract build method logic to BuildMethod classes
Rather than having various build method-specific functions within
`Build`, and switching which of these to execute in `Build` based on the
`build_method` specified for each node, this commit extracts a
`BuildMethod` instance for each node and has `Build` just call methods
on this to have it render the appropriate templates.

This simplifies the code in `Build` and should make things more easily
extensible for adding future build methods or expanding the existing
ones.

As suggested in
https://github.com/alces-software/metalware/pull/156#pullrequestreview-55560842.