Extract specifying template paths to BuildMethods
Seems sensible to define the method for getting the build template paths
for a node in the same place as we now render those templates.

While extracting these I also noticed that we were specifying a
dependency on the `pxelinux` and `kickstart` templates even when
building a node using the `basic` build method, so this has now been
corrected; we don't yet specify a dependency on the required `basic`
dependency in this case, but this can be added at some point in the
future.