Remove Kickstart-specific language from Build
We now support two different 'build methods', and may support more in
future, so we should rename the methods in Build with names specific to
a Kickstart build.

I've also tweaked the messages printed when we ask if we should
re-render the PXELINUX configs, to be less Kickstart build-specific - we
may or may not want to keep this change, as as it stands Kickstart built
nodes remain the only case where anything will actually be re-rendered.