Add test for increasing value of index
The group cache index is incremented for every index added, and is not
decremented even if a group is added and then immediately removed.

This test documents this behaviour and demonstrates why we currently
need to track the `next_index` in the cache as well as the current map
of groups to indexes (if we did not need this behaviour we could just
work out what the next index should be by incrementing the current
maximum index).