Improve error reporting of UnsetParameterAccessError
Previously if a unset parameter was detected, a new error object was
created and raised. However this obfuscates where the error is actually
comming from.

Now a new error object is still created, however its backtrace is set
to equal that of the old error. This gives much greater detail on
where the error can from