Check config for DNS type in DomainTempatesRenderer
The domain templates renderer needs to check if it should use the
hosts file or named for DNS. The easiest way to implement this is
to check if the value has been set in the domain.yaml config file.

By default,if the value hs not been set, it will issue a warning and
then use 'hosts' DNS. This is to prevent the existing tests from
breaking.

The dns_type can then be linked to a question in 'configure.yaml'.