Implement basic logic for named DNS
Created a DNS module to contain the Named code. The hosts code can
be moved into this module at a latter date.

During DomainTemplatesRenderer, instead of just rendering the hosts
file, the dns_type is checked from the rendered repo_config for
just the domain. This means if dns_type is changed or dependant on
a group or node answer, it will not be picked up. However this
decision was made as the dns type SHOULD be set first with configure
domain and then remain static from then on.

The code within `DNS::Named` is for the basic setup and testing. If
the named server is not running, then it goes through the setup,
otherwise it will continue with the `update` method which hasn't been
fully implemented.

Also updated the install script to download the required named modules.
It is assumed that named will be used most of the time so it is
probably safe to always download the modele.