Render /etc/named.conf during named setup
Renders the template saved in scripts/templates during setup. The
rendering is reasonable straight forward but needed to add some
logic so that the files where avaliable in the fake filesystem.

The actual template being used is spoofed for the testing however
the paths are correct. This is to make testing the ruby component
easier.