Restart named on DNS::Named.setup
Runs the required `systemctl` commands when setting up named. Also
added logging to named setup.

The majority of the code changes in this commit is adding safety to
the tests to completely stub the restart_named method. This is because
the system commands could be dangeours if they are accidentally ran.