Rerender the metalware named conf on update
The repo specific named template needs to be rerendered each time.
The template file path is pulled from `domain.yaml` and thus the
file_path method needed to be updated so it works without a node
specified.

The zone and net information for the network are inputed under a
alces_named key. Ideally this would be moved under the alces namespace