Return raw config values from Binding
Allow the BindingWrapper to pull values from the raw config. This is
done by assuming any value that is not in the alces namespace is a
config value. The alces magic namespace has not been implemented at
this stage.

Then when returning the results, if it is meant to return a hash
containing more config values, it actually returns itself. The method
call is then added to the bindings call stack. This call stack is
repeated to get any nested values.

When returning to the templater, the results are put into a new
wrapper containing the updated call stack. The only exception is
TrueClass, FalseClass and the to_s method which returnt the results
directly.