Return and evaluate through the actual binding
The purpose of the BindingWrapper is to act as the binding for ERB.
This means that instead of calling methods on the object directly,
the scope of BindingWrapper is returned as a binding.

Also as the BindingWrapper needs to be removed before printing to the
template, the to_s method must not be returned with a wrapper. As
their maybe other edge cases which require this, all `to_` methods
return without the wrapper.