Implement saving of configure domain answers
This runs the entered answers through the configure command class
directly, so that everything which would happen if this command was run
on the command line should still occur.

This commit introduces a generic way GUI controllers can run Metalware
commands and have any error message produced appear in the GUI via a
flash message.