Allow recursive ERB in config
Now that config values are rendered when they are needed, it forms
a recursion to replace embeded ERB. This means that the deepest
level of imbeded erb is replaced first before being subsituted into
the higher levels.

This removes the need to rerender the config until it remains the
same; as was previously the case. This does make it easier for a
infinite recursive loop to form, which is why a loop counter has been
added.