Allow maths operations on the binding
Previously the returned value from the Binding was always parsed for
ERB. This however causes an error as Integer data types are not a
valid ERB template. Now only Strings are templated as they are the
only values that can contain ERB.

Also as maths opertations can not be performed on the Binding class
directly, a coerce method had to be added. This coerce method returns
the wrapped object (alces_binding).