Log and return missing config values as nil
When a missing config value is detected, the callstack is logged as
a warning and a Nil is returned. This Nil is not contained in the
wrapper so it is a falsy value.

Also prints the call stack when any StandardError has occured when
trying to retrieve a value.

Also added a test to make sure that the configs are being merged for
a node. This was already the case due to Templating::Configurator,
however the test confirms it.