Only configure with entered answers from form
Without doing this we would configure with all the answers, with an
empty string for those without any input entered. This is inconsistent
with the CLI behaviour and would mean that we would not follow the
hierarchy when configure data is later loaded (as every single answer
would always be saved at every level).