Allow the binding to be directly queried for a value
Previously the Parameter binding only worked within the Wrapper thus
making it tricky to directly query from a value.

Now it possible to create a Parameter object without the outer wrapper
and then directly query for a value. Currently the implementation
is a bit strange as Parameter is highly coupled with Wrapper and thus
the inputs needed spoofing. This should can be addressed at a future
date.