Extract defining Configurators to own class
Previously we created the `Configurator` to use in `ConfigureCommand`,
by calling a couple of methods which should be defined in sub classes
along with inspecting the name of the command to determine the questions
section to load.

This commit changes this so each configure command explicitly calls an
appropriate `Configurator` class method to create the `Configurator` it
should use. This is both clearer and means we can easily create the same
`Configurator`s outside of `ConfigureCommand`s, which will be useful now
in the new GUI.