Use questions from Configurator in configure form
Rather than loading the questions data from the `configure.yaml` file
directly, this commit changes this to use the new
`Configurator.for_domain` method to get a `Configurator` with access to
the fully parsed `Question` objects.

This is useful as it will allow us to use other aspects of the parsed
`Question`s in the configure form, such as the default value.