Create the MagicNamespace inside the Binding
As the binding relies on the MagicNamespace existing, it should be
responsible for creating it. Thefore the additional parameters
required to create the namespace are passed into the Binding and then
the namespace is created.

Also the Node object used by the Namespace is now created internally
and thus only the node name as a string is required as an input.