Create a new Binding inside GroupNamespace
When iterating through the nodes  in the GroupNamespace, the config
specific to each node is used. This requires returning a new binding
for each node

NOTE: it does not share the same Parameter object because the Node
in theory could overide any config parameter. This means the config
parameters will need to be rerendered for each node, hence why they
get a new binding