Remove MissingParameterWrapper
The MissingParameterWrapper is now been replaced with Binding::Wrapper.
This commit removes referecens to the older wrapper. The new wrapper
will automatically wrap any return objects, so their is no need to
manually wrap the answers anymore.

This does mean that the test for wrapper had to be removed as it nolonger
made sense it that context