Recursive render Answer ERB hack
This commit is a bit of a hack around the infinite recursion caused
by the answers. It is not possible to render the answers in full
whilst only pulling Config values as needed.

(Well actually it is, but it is SUPER DOOPER easy to form an infinite
loop)

So instead if the Answer contains ERB, that ERB is rendered into
the template. Then if the template still contains ERB, then it is
rerendered against the same binding until it is removed