Split build command with edit flags
The build command can be inturrepted after the templates have been
rendered with `--edit-start` command. This prevents `build` from
continuing with the build process, allowing the files to be edited.

Then the build process can be continued with `--edit-continue` which
will skip the render process and directly wait for the built nodes.

The two flags can not be used together as that does not make sense.