Add file autocompletion to edit command
A new autocomplete key in the CLI's config.yaml includes a list of
values to add to the autocompletion. If however the value starts
with COMP_FILE::, then bash autocompletion will switch to autocomplete
the filename. I couldn't get it running with relative paths, so
the absolute path is used instead