Add JavaScript runtime dependency so Rails works
For some reason this was needed after setting up a new development VM
when it wasn't before. Possibly this is due to the previous VM having
node installed (for flight-radar?), but this cannot easily be installed
now due to https://bugzilla.redhat.com/show_bug.cgi?id=1481470. So maybe
we can remove this dependency if that is fixed. In any case, this
appears to fix the issue for now.