Restart named after each update
Named needs to be restarted when ever an update command is ran, not
only when it is started.

The methods have been slightly renamed. During setup, start_named is
called to run the bash commands which end with restart_named. But if
named, has already been setup, then restart_named is called only.

The test have been updated to reflect this