Suppress missing nodename warning
Because of how the config is being rendered, the nodename is being
accessed when a node hasn't been set. However this is an artifact
of the rendering of the Config and not a legitimate warning.

Also in principle, if you render for a node, nodename would be set
and if you're rendering for the domain, you shouldn't need nodename.
However due to the complexity of the config rendering, this can not
be insured.

This commit suppresses the warning as for the most part it is erroneous.
However their maybe case where it should be issued but isn't. To fix
this, additional work is required on the namespace