Base UEFI build of Kickstart with diff template
Created the UEFI build off the kickstart build. The differance
between the two builds has not yet been implemented and therefore
the UEFI class is currently empty. This will change in future
commits.

The major change is setting the default pxelinux template for a
UEFI build using the new DefaultFile class. This is tested using
the private method as to avoid editting the BuildMethod::BuildMethod
class.