Confirm merge order for node configs
Tests that the 'self' node access the same configs. At this point in
time, it behaves as a regular node. This means that the self config
will inherit domain level configs. However no other node will inherit
the self config.

This does meant that if self is added to additional groups, then the
group level configs will also be merged in. However as this is some
what expected behaviour, it will remain for now. This can be revisted
at a future date if it causes problems.