Add tests for correct build type for self node
As the self node can only be built with the self build type (which
will be implemented in a further commit), this commit adds the tests
for it.

Once implemented, the self build will error if the build type is
specified to be anything other than self. If the build type is not
specified then it will implicitly use it.