Return the special build type for self node
As the self node can only have the self build type, special logic is
required to enforce this. An error will be raised if the build type
is not self.

Also if another node tries to use the self build type, this will also
raise an error.

Also implemented the self build method off the basic build. The only
difference is the self build uses the self template. Note the tests
will fail as self build method isn't being required. This will be
fixed in the next commit