Specifiy the start build command in BuildMethods
Now that their are other build types that are not pxelinux based,
a command maybe required inorder to start the build process. For this
reason, a start_build method in BuildMethod will be ran after the
templates have been rendered. By default, it is blank and does nothing.

In the case of the self build, the self template is ran using bash