Allow recording detected nodes on hunter page
A form is now provided for each detected MAC address in the table, which
allows the user to associate a name with this MAC address in the
`hunter.yaml` file.

Note: the submit button for each form is associated with the form using
the `form` attribute, which apparently won't currently work in IE/Edge;
we may or may not care.