Create base class for validating save files
Similar to the Loader class, the Saver clas will validate files before
saving them. This is to prevent invalid files from being saved, which
makes it hard to correct as the original file has been over ridden.

The first input to all save methods must be the data to be saved. This
save data will then be avaliable in the save method as the data method.