Display any warnings produced on build page
When we would previously display a warning to stderr, we now switch on
whether we are in the GUI or not (determined by whether `Rails` is
defined); if we are, each warning message is then stored locally to the
current thread. The stored messages are then displayed in order on the
build page.

Other message types will be handled similarly in future commits.

The behaviour when we are in the CLI should be retained as before.