Keep complete build command/thread alive when in GUI
We want to be able to view the status of/messages from a build after it
has completed, otherwise these will disappear too soon and the build
page would appear the same as if the build  had never happened.

Keeping the build thread, which holds all the messages, alive once the
build is complete is the simplest way to do this for now, though this is
not very clean or efficient and we should change this sooner or later.