Gracefully shutdown build commands run from GUI
Rather than just killing the build thread, which means the build
complete (i.e. PXELINUX) templates won't be re-rendered and all messages
output by the build command are immediately lost, we instead now
gracefully shutdown the build command; this means the build complete
templates will be re-rendered and we keep around the build messages to
continue displaying.

This is done by setting a flag in the build command thread to tell it to
gracefully shutdown; this flag is then repeatedly checked by the command
as the first task in its loop to check for built nodes and, if set, the
command will appropriately shut itself down and sleep indefinitely.