Support cleaning up completed build command
Now that we gracefully shutdown a running build command, to perform the
necessary actions when a build is complete and retain access to the
build messages, we also need to provide a way to kill the completed
build thread outright so that the build state is reset and a new build
can be started if needed.

We now record if a build is complete (all nodes have built or the
command has been gracefully shutdown) within the command to a thread
variable; this is then checked in the view to determine which button to
show, and a completed build thread can thereby be killed.