Implement polling for newly detected MAC addresses
When we are hunting, every second we now poll for and append table rows
for newly detected MAC addresses to the hunter page's detected nodes
table. This means that once a user has started the hunter process they
no longer need to refresh the page to see new data appear.

This polling is handled using the Intercooler JavaScript library
(http://intercoolerjs.org/).