Implement validation on choice questions
Instead of implementing choice as a question type, it is actually
an additional parameter to the question. This allows the type of choice
expected to be set independantly.

If choice is to be used with a default, the default must appear in
the choice list. All choices must match the question type