Refactor so only need single build show view
Previously we had different `show` views for building nodes and groups,
which both just rendered the same partial with certain paths, specific
to the type of thing being built, passed in.

By inspecting the name of the build controller we can dynamically
determine the type of item being built, and can then use this to ensure
the correct path methods for this type are defined and used within the
view.

By doing this we can just have a single view for `Build#show` used in
both cases, similarly to what we already have for `Configure#show`. This
means we no longer need to re-specify similar paths in multiple places,
which will be particularly useful when adding additional views which
should also share the same paths.