Return all object in the MissingWrapper
Inorder to track the call stack, all objects returned from the
MissingParameterWrapper must also be wrapped with exceptions. Any
falsey values (false, nil) can not be wrapped otherwise they become
truthy. The TrueClass is also not wrapped for consistancy.

Also as ERB needs to convert objects to strings before placing them
in the template, the `to_s` method can not return a wrapped object.
As their are other use cases where this is the case, all the to_*
methods do not return a wrapped object.

The MissingParameterWrapper has also been made a Blank class and as
had a binding method put onto it.