Track the callstack as part of the MissingWrapper
The MissingParameterWrapper now keeps track of the callstack so it
can print which call returned a nil.

Some additional code was required to handle array indexing. One of the
objects received the index method as an integer which didn't previously
cause an issue as it wasn't wrapped. The fix is to make sure the integer
is called with the index method: '[]'