Ensure GUI aware build has completed if it errors
We should handle a build which has finished due to the build command
erroring in the same way as a build which has finished in a more
expected way; i.e. by providing the option to clean this build up (kill
the build thread) rather than what we did previously, provide the option
to cancel the build (attempt to gracefully shutdown the build command),
which would never be picked up as the build command has already exited.