Fix inheriting of higher level false boolean answers
Since false and nil are both falsy, the way we were determining the
question default for boolean questions with a false answer saved in a
higher level answer file was slightly broken, as the false answer would
be ignored in the same way as a non-existent/nil answer.

This commit fixes this so we explicitly check whether the answer defined
at a higher level is nil, and only fall back to using the
`configure.yaml`-defined default as the question default in this case.