Improve determining whether boolean questions enabled
Previously we did not handle what the initial state for boolean
configure questions should be in various situations, so they would
default to false/'No' in several cases where there would be a more
logical/expected initial state. This commit gives more consideration to
this; see inline comments for explanation.