Define templates list of repo directory
Between the kickstarts methods, the only information that changes is
the save paths and the repo directory which to find the template.

The directory is now set by REPO_DIR which means the pxelinux_repo_dir
and TEMPLATES can be set in the base Kickstart class. This did require
setting the TEMPLATES on initialize otherwise the Kickstart class
itself will try and set TEMPLATES on require which causes an error.