Inline handling running configure in controller
When I originally added `ApplicationController#run_command` I thought
how we would run Metalware commands within the GUI would be more
generalizable than it turned out to be. Since this method is only used
in one place I'm inlining it for now; we can always generalize this in a
better way in future if possible.