Use the Loader in Configurator for configure_data
This commit switches the Configurator over to using the Loader method.
This means the file is validated on load automatically.

So the tests do not need to be completely rewrote, the Loader method
has been stubbed to return the data. This means Validation does not
occur in these tests by will occur in production.