Use Loader/ Saver for answer files in Configurator
This commit switches the configurator over to use the Loader and Saver
for managing the answer files. This way validation automatically occurs
on these files.

Also the answer file path no longer needs to be an input, only the
name of the node or the group

Also switched the section being tested to :domain as a valid section
now needs to be used inorder for the saving to take place.