Fixed test to make sense with Validation
Now that their is validation of answers, it is not possible to save
an answer file with a question answer that doesn't exist. This means
the previous test didn't make sense as it had an answer to a missing
question and thus errored.

The test has been updated so that all questions have been defined and
no question receive a new answer.