Spoof the Loader in Repo tests
The hash format is the correct format to be returned by the Loader
however it is the incorrect format for Validation. This is because
the validator converts to a hash as its last step.

It is therefore easy to spoof the loader to directly return the hash.
Otherwise the configure.yaml needs to be saved in array format and then
converted to a hash by the validator