Spoof configure_data loader in Node tests
The hash format for configure_data is the correct return value. However
the validator expects it to be in array format for validation. The
validator would then convert to an hash format at the end except it is
failing before that.

This commit spoofs the loader to directly return the hash instead.