Change Configurator.for_node input to node_name
Previously a Node object was passed in which made it inconsitant with
for_group which took the group_name. This commit changes it to taking
the node_name instead.

This does cause a test to fail as the primary group is no longer known
for the node_name. But this will be fixed when higher_level_answer_files
is removed as an input from Configurator