Fix configure self to use self answers
With the new refactoring the self questions where not being used for
configure self. Instead the self node was being treated as a regular
node. This commit sets up the path methods to retreive the self answers
through the file loader.

The Configurator also know supports an empty hash of questions, where
previously it would error.

The self answers have been moved as they aren't the same as a regular
node and therefore shouldn't be saved as such. The self answers are
now stored along with domain.yaml as self.yaml