Correctly handle array indexing in Parameter Wrapper
Originally the array index method (`[]`) was required as the
method_missing didn't pass in all the arguments. However this then
created bugs which additional code was introduced to patch.

Turns out that all the patches/ code changes mentioned above is now
superfluous as the method_missing `send(s,*a,&b)` can natively handle
the array index methods.

NOTE: The array index method `[]` is also used for hash indexing which
is required by ERB.