Fix answer loading in Configuration
The configuration class use to build it's own answer paths and load
the files directly. This skips the answer validation and means it
looked for the self answers in the wrong place. The Configuration
class still needs work as it will break if a node is called any of
the reserved words: domain, group, node, self

As the Loader class is now used, the answers validatation needed to
be turned off for a few tests as they are legacy tests that do not
set up configure.yaml correctly