Transform JSON answers to hash with symbol keys
The --answer input allows the configurator to take a JSON string and
use it as the answers. However the validation that occurs on it,
expects the keys to be symbols. This causes the validation to break.

The fix is to convert the JSON to hash and then to symbolize all the
keys.

Fixes #217