Load the configs through the MergedHash
This commit creates a new utility MergedHash, it will be responsible
for loading the config and answers. Eventually this class will be
embeded into the Namespace.

The key idea behind it is it loads the answers and config in the same
way. Currently however only loading the domain config is supported.