Load answer through the MergedHash
As the sequence to merge the Answer hashes is the same as the config,
the majority of the code is the same.

The only major difference is to the `load_yaml` method which must
vary to use the `loader`.

Also as part of the testing, sometimes it is ideal to skip file validation
as that is not what the aim of the test is. To facilitate this, a
validation field has been added to Metalware::Config. When set to
false it will cause the answer file validation to be skipped. Currently
no other validation is skipped but this can be added latter if
required.