Create MetelRecursiveOpenStruct
Metalware requires specific behaviour for it's RecursiveOpenStruct
and thus a new 'Metal' version has been created.

It will eventually be merged with IterableRecursiveOpenStruct however
the two code bases are beeing kept seperate for the time being. This
is to prevent breaking Metalware in its current version.

The other function of MetalRecursiveOpenStruct is to render the strings
on the fly. This means it will need the AlcesNamespace inorder to
template it's parameters. The alces parameter is passd in and stored
however it is not included in the Hash.