Recursively replace ERB strings in hash
The MetalRecursiveOpenStruct now renders all string values before
returning. This means a config value should never be returned whilst
containing ERB (*exception).

The templater itself is part of the Alces namespace as it acts as the
binding (*this may change with the Wrapper). The Alces namespace is
responsible for keeping track of the recusion depth.

Exception:
It maybe possible to escape ERB which will return regular ERB. If this
ERB doesn't go through the templater again, then it won't be replaced.
It should be possible to preserve the escapped ERB by re-escapping this
ERB. However this can be addressed in the future.